requirements.txt
# Core Django
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Database
psycopg2-binary==2.9.11

# Media handling
pillow==11.2.1

# Configuration
python-decouple==3.8

# Production dependencies (S3 storage - optional)
boto3==1.38.46
botocore==1.38.46
django-storages==1.14.6
s3transfer==0.13.0

# ASGI server (optional)
uvicorn==0.34.3

# Dependencies of above packages
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
h11==0.16.0
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
